refactor: moved information screen presenter
