+++ Sources/Utilities/OnboardingViewControllerFactory.swift
@@ -10,4 +10,12 @@ final class OnboardingViewControllerFactory {
         }
         return IntroViewController(viewModel: viewModel)
     }
+    
+    static func createPasscodeInformationScreen(analyticsService: AnalyticsService,
+                                                action: @escaping () -> Void) -> GDSInformationController {
+        let viewModel = PasscodeInformationViewModel(analyticsService: analyticsService) {
+            action()
+        }
+        return GDSInformationController(viewModel: viewModel)
+    }
 }
